Countdown template added
Countdown template was added. There is a problem with setting the
countdown.
Javascript path is: countdown\js\simplyCountdown.js
I believe countdown can be set on line 118, 119 and 120. But changes
don't work. Need help.
Colors and fonts must be checked if there is a need for changing them.
Otherwise they're good to go.
e-mail send button not working yet, will try to activate it soon.
Template seems to not work on IE, will analyze the reasons.